65: Updating Postgres container version to match changes being made to prod.
Updating Postgres from v9 -> v11 requires a couple of changes around the
earthdistance extension.  Data can't be reloaded because of a mismatch
in extension versions so an index needs to be dropped before migration
and re-instated after.  Future backups should load fine, but currently
anything we download from S3 can be used with caveats...

- Update Docker container to Postgres 11
- Add creating of extensions to setup
- Split untar and restore in the restore script allowing us to work
with existing directories
- Add some notes in the restore script
- Change order so that dojos (which may fail) goes last
- Remove insert for address column that no longer exists from anon sql